Refactor: rename UseInput interface
