+++ src/interfaces/basic.ts
@@ -1,6 +1,6 @@
 import { ChangeEventHandler, MouseEventHandler } from 'react';
 
-export interface IUseInput {
+export interface UseInput {
   value: string;
   placeholder: string;
   onChange: ChangeEventHandler<HTMLInputElement> | undefined;
